server side done.

+++ tower/server/services/TicketsService.js
@@ -8,10 +8,19 @@ class TicketsService {
     return tickets
   }
   async deleteTicket(ticketId, userId) {
-    const ticket = await dbContext.Tickets.findById(ticketId).populate('ticket profile')
+    const ticket = await dbContext.Tickets.findById(ticketId)
     if (!ticket) throw new BadRequest(`Ticket at id: ${ticketId} does not exist`)
+    if (ticket.accountId != userId) throw new BadRequest("You can't keep me out of the chocolate factory")
     await ticket.delete()
-    return `ticket for ${ticket.profile.name} to ${ticket.event.title} is destroyed`
+    return `ticket is destroyed`
+  }
+  async findAllTickets(eventId) {
+    const allTickets = await dbContext.Tickets.find({ eventId: eventId }).populate('profile event')
+    return allTickets
+  }
+  async findMyTickets(userId) {
+    const myTickets = await dbContext.Tickets.find({ accountId: userId }).populate('event')
+    return myTickets
   }
 }
